Add a Make target to scrypt-1.2.0 for printing vars.
Allows us to export local Makefile variables to upstream make;
will come in extremely handy when trying to build scrypt as a
library.

The drawback is this modifies the scrypt Makefile.am itself, which
means future scrypt upgrades won't be drop-in unless the print
target is re-added to the distribution.

Then again, the planned approach for compiling scrypt as a library
isn't foolproof against major upgrades (such as1.1.6 to 1.2), as well
as releases generated with a substantially different version of
autotools.